+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 3.9.2 (2015-xx-xx)
 --
 Improvements:
-* Core: every template can now check if it has a certain parent id with {option:isChildOfPage1}
-
+* Core: every template can now check if it has a certain parent id with {option:isChildOfPageX}
 * Locale: improve performance of the index page
 * Core: reduced database queries in BackendDataGridFunctions::getUser().
+* Core: Replace SITE_MULTILANGUAGE with $container->getParameter('site.multilanguage')
 
 Bugfixes: